<feat> [be] : 녹강 join을 email -> id

+++ backend/src/main/java/com/yoga/backend/mypage/recorded/repository/MyLikeLectureListRepository.java
@@ -20,7 +20,7 @@ public class MyLikeLectureListRepository {
         this.queryFactory = queryFactory;
     }
 
-    public List<LectureDto> findMyLikedLectures(Long userId) {
+    public List<LectureDto> findMyLikedLectures(int userId) {
         QRecordedLecture lecture = QRecordedLecture.recordedLecture;
         QRecordedLectureLike like = QRecordedLectureLike.recordedLectureLike;
